pyproject.toml
[project]
name = "tangle-detect"
version = "0.1.0"
description = "Agent workflow deadlock and livelock detection"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Design Team" },
]
keywords = [
    "agents", "langgraph", "deadlock", "livelock",
    "monitoring", "observability", "opentelemetry", "otel", "otlp",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Monitoring",
]

readme = "README.md"

dependencies = [
    "xxhash>=3.5.0",
    "pydantic>=2.10",
    "structlog>=24.4",
]

[project.optional-dependencies]
langgraph = [
    "langgraph>=0.3",
    "langchain-core>=0.3",
]
server = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
]
otel = [
    "opentelemetry-api>=1.29",
    "opentelemetry-sdk>=1.29",
    "opentelemetry-exporter-otlp-proto-grpc>=1.29",
]
metrics = [
    "prometheus-client>=0.21",
]

[project.urls]
Homepage = "https://github.com/nobelk/tangle"
Repository = "https://github.com/nobelk/tangle"
Changelog = "https://github.com/nobelk/tangle/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/nobelk/tangle/issues"

[project.scripts]
tangle = "tangle.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tangle"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "/tests",
    "/examples",
    "/Dockerfile*",
    "/docker-compose.yml",
    "/Makefile",
    "/.python-version",
    "/uv.lock",
    "/tangle-python-design-doc-v3.md",
    "/CLAUDE.md",
    "/COMPATIBILITY.md",
    "/.claude",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "pytest-timeout>=2.3",
    "pytest-benchmark>=4.0",
    "hypothesis>=6.120",
    "ruff>=0.8",
    "mypy>=1.14",
    "httpx>=0.28",
    "prometheus-client>=0.21",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "grpcio>=1.68",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
timeout = 30
addopts = "-ra --strict-markers"
markers = [
    "integration: marks tests that require external services",
    "slow: marks tests that take >5 seconds",
]
